+++ CHANGELOG.md
@@ -1,11 +1,26 @@
 # OpenStudio Extension Gem
 
+## Version 0.2.1
+
+* Changes from 0.1.5 (runner.conf bug)
+* Changes from 0.1.6 (update core library)
+* Update measure tester to 0.2.2 (with Rubocop 0.54)
+
 ## Version 0.2.0
 
 * Upgrade Bundler to 2.1.x
 * Restrict to Ruby ~> 2.5.0
 * Remove json_pure gem
 * Update measure tester to 0.2.0 (removes need for github checkout)
+* Note that this version does not include the changes from Version 0.1.5
+
+## Version 0.1.6
+
+* Update core library methods with what was in OpenStudio-measures
+
+## Version 0.1.5
+
+* Fix bug to respect the runner.conf file when using Extension::Runner with no options
 
 ## Version 0.1.4